+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## v1.15.0
 - Add option to calibrate using the unique device ID instead of the standard device ID (useful for arduino's and other similar devices)
+- Attempt to register joydevice on the fly in case the joy device added event is not processed/fired
 
 ## v1.14.3
 - Add notice about Cab Debugger functionality