+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add option to calibrate using the unique device ID instead of the standard device ID (useful for arduino's and other similar devices)
 - Attempt to register joydevice on the fly in case the joy device added event is not processed/fired
 - Improve direct control server restart handling and error logging
+- Website security updates
 
 ## v1.14.3
 - Add notice about Cab Debugger functionality